Fix signed integer overflow in CbcNode trust-adjustment check
CbcNode::chooseDynamicBranch() computed
'model->numberStrongIterations() * 20 < model->getIterationCount()' to
decide whether to increase trust in pseudocosts. On long-running
solves numberStrongIterations() can exceed INT_MAX/20, so the '* 20'
overflows a 32-bit int (UBSan: signed integer overflow: 107374237 * 20
cannot be represented in type 'int'), observed once in a recent UBSan
sanity run.

Promote the left-hand side to double before multiplying so the
threshold comparison can't overflow; the comparison only needs to be a
threshold check, not exact integer arithmetic. Verified with the full
Cbc mip-sanity-data suite (473/473 passed, no regressions).

+++ src/CbcNode.cpp
@@ -4873,7 +4873,7 @@ int CbcNode::chooseDynamicBranch(CbcModel *model, CbcNode *lastNode,
       model->setSearchStrategy(2);
     }
   }
-  if (numberUnfinished * 10 < numberStrongDone && model->numberStrongIterations() * 20 < model->getIterationCount() && !auxiliaryInfo->solutionAddsCuts()) {
+  if (numberUnfinished * 10 < numberStrongDone && static_cast< double >(model->numberStrongIterations()) * 20 < model->getIterationCount() && !auxiliaryInfo->solutionAddsCuts()) {
     //printf("increasing trust\n");
     model->synchronizeNumberBeforeTrust(2);
   }
